Restructure of Endpoints
